Revert "saving, need to fix pdfs"
This reverts commit b2dcaac9a29ec1a717f8bddfd2d3a3aa50283069.

+++ src/sampleShapes.h
@@ -24,8 +24,8 @@ __host__ __device__ ShadeableIntersection sampleSphere(Vector2f &xi, float& pdf,
 	it.surfaceNormal = glm::normalize(it.intersectPoint - lightCenter);
 
 	float dist = glm::distance(refPoint, lightCenter);
-	float sinThetaMax2 = 1.0f / dist*dist;
-	float cosThetaMax = std::sqrt(glm::max(0.0f, 1.0f - sinThetaMax2));
+	float sinThetaMax2 = 1 / dist*dist;
+	float cosThetaMax = std::sqrt(glm::max((float)0.0f, 1.0f - sinThetaMax2));
 	pdf = 1.0f / (2.0f * PI * (1.0f - cosThetaMax));
 
 	return it;
